unnecessary component removed
