--- src/components/molecules/ProtocolsIcons/styles.scss
@@ -1,5 +1,0 @@
-.protocols-icons {
-  & img:not(:first-child) {
-    margin-left: -4px;
-  }
-}
