+++ src/pages/Txs/index.tsx
@@ -3,12 +3,13 @@ import { useSearchParams } from "react-router-dom";
 import { useQuery } from "react-query";
 import { useEnvironment } from "src/context/EnvironmentContext";
 import { BlockchainIcon, NavLink, Tooltip } from "src/components/atoms";
-import { CopyToClipboard, ProtocolsIcons, StatusBadge } from "src/components/molecules";
+import { CopyToClipboard, StatusBadge } from "src/components/molecules";
 import { SearchNotFound } from "src/components/organisms";
 import { BaseLayout } from "src/layouts/BaseLayout";
-import { parseAddress, parseTx, shortAddress } from "src/utils/crypto";
+import { filterAppIds, formatAppIds, parseAddress, parseTx, shortAddress } from "src/utils/crypto";
 import { timeAgo } from "src/utils/date";
 import { ArrowRightIcon, CopyIcon } from "src/icons/generic";
+import { allBridgeIcon, cctpIcon, mayanIcon, nttIcon, portalIcon } from "src/icons/protocols";
 import { getChainName, getExplorerLink } from "src/utils/wormhole";
 import { ChainId, ChainLimit, Order } from "src/api";
 import { getClient } from "src/api/Client";
@@ -16,10 +17,12 @@ import { GetOperationsInput, GetOperationsOutput } from "src/api/guardian-networ
 import { Information } from "./Information";
 import analytics from "src/analytics";
 import {
+  ALL_BRIDGE_APP_ID,
   CCTP_APP_ID,
   CCTP_MANUAL_APP_ID,
   CONNECT_APP_ID,
   IStatus,
+  MAYAN_APP_ID,
   NTT_APP_ID,
   PORTAL_APP_ID,
   UNKNOWN_APP_ID,
@@ -48,6 +51,13 @@ export const PAGE_SIZE = 50;
 export const ETH_LIMIT = {
   maxTransactionSize: 5000000,
   availableNotional: 50000000,
+};
+
+const protocolIcons: Record<string, string> = {
+  [ALL_BRIDGE_APP_ID]: allBridgeIcon,
+  [CCTP_APP_ID]: cctpIcon,
+  [MAYAN_APP_ID]: mayanIcon,
+  [NTT_APP_ID]: nttIcon,
 };
 
 const Txs = () => {
@@ -313,6 +323,22 @@ const Txs = () => {
                 }
               }
 
+              let portalDisplayed = false;
+
+              const appIdsToDisplay =
+                appIds?.length > 0
+                  ? filterAppIds(appIds)
+                      .map(appId => {
+                        const iconSrc = protocolIcons[appId] || portalIcon;
+                        if (iconSrc === portalIcon && portalDisplayed) {
+                          return null;
+                        }
+                        portalDisplayed = iconSrc === portalIcon;
+                        return iconSrc;
+                      })
+                      .filter(Boolean)
+                  : [];
+
               const timestampDate = new Date(timestamp);
               const row = {
                 VAAId: VAAId,
@@ -457,7 +483,19 @@ const Txs = () => {
                     <h4>PROTOCOL</h4>
 
                     {appIds?.length > 0 ? (
-                      <ProtocolsIcons appIds={appIds} />
+                      <div className="tx-protocols-icons">
+                        <Tooltip
+                          maxWidth={false}
+                          tooltip={<div>{formatAppIds(appIds)}</div>}
+                          type="info"
+                        >
+                          <div>
+                            {appIdsToDisplay.map(icon => (
+                              <img key={txHash} src={icon} alt={icon} height={24} width={24} />
+                            ))}
+                          </div>
+                        </Tooltip>
+                      </div>
                     ) : (
                       <div className="not-found">-</div>
                     )}
